Relocate password input screen to a separate file
